+++ CHANGELOG.md
@@ -5,6 +5,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.6.3] - 2023-10-31
+
+### Added
+
+- Enable computation of curve torsion (`TorsionScalar`).
+  This is only possible with high-order filament discretisations (`QuinticSplineMethod`, `FourierMethod`).
+
 ## [0.6.2] - 2023-10-30
 
 ### Changed